requirements.txt
gspread==6.1.2
google-auth==2.34.0
python-dateutil==2.9.0.post0
typer==0.12.5
rich==13.8.1
fastapi==0.114.1
uvicorn==0.30.6
jinja2==3.1.4

